+++ tests/Logic/StatementTest/StatementCreatorTest.php
@@ -108,6 +108,7 @@ class StatementCreatorTest extends TestCase
             $this->createMock(CommonHelpers::class),
             $reusableMessageBlocks,
             $this->createMock(XBeteiligungAuditService::class),
+            new Din91379TextSanitizerService($this->createMock(LoggerInterface::class)),
             $this->createMock(XBeteiligungPhaseDefinitionCodeRepository::class),
         );
         $this->XBeteiligungService = $xbeteiligungService;